move Today style files to style/ subfolder
