+++ client/src/components/Today/Note.tsx
@@ -3,8 +3,8 @@ import type { NoteWithIds } from "@type/server/note.types";
 import type { TagWithIds } from "@type/server/tag.types";
 import type { ById } from "@type/server/utility.types";
 import TagCard from "../TagCard/TagCard";
-import N from "./Notes.style";
-import S from "./Today.style";
+import N from "./style/Notes.style";
+import S from "./style/Today.style";
 
 type NoteProps = {
 	note: NoteWithIds;
