+++ client/src/components/Today/style/AllDayActivity.style.ts
