+++ client/src/components/Today/style/DetailedActivity.style.tsx
