+++ client/src/components/Today/style/Notes.style.ts
